Modified max value of maxgeocells

+++ warp10/src/main/java/io/warp10/script/MemoryWarpScriptStack.java
@@ -247,7 +247,8 @@ public class MemoryWarpScriptStack implements WarpScriptStack, Progressable {
     setAttribute(WarpScriptStack.ATTRIBUTE_MAX_PIXELS, Long.MAX_VALUE - 1);
     setAttribute(WarpScriptStack.ATTRIBUTE_URLFETCH_LIMIT, Long.MAX_VALUE - 1);
     setAttribute(WarpScriptStack.ATTRIBUTE_URLFETCH_MAXSIZE, Long.MAX_VALUE - 1);
-    setAttribute(WarpScriptStack.ATTRIBUTE_MAX_GEOCELLS, Long.MAX_VALUE - 1);
+    // Set max of geocells to the largest INTEGER - 1, not Long.MAX_VALUE as it is used as an int
+    setAttribute(WarpScriptStack.ATTRIBUTE_MAX_GEOCELLS, Integer.MAX_VALUE - 1);
   }
   
   @Override
@@ -1113,7 +1114,6 @@ public class MemoryWarpScriptStack implements WarpScriptStack, Progressable {
   
   @Override
   public Object setAttribute(String key, Object value) {
-    
     if (null == value) {
       return this.attributes.remove(key);
     }
